Coded All selfProductService Controllers

+++ src/main/java/com/scaler/productservice/controllers/ProductController.java
@@ -2,12 +2,12 @@ package com.scaler.productservice.controllers;
 
 import com.scaler.productservice.exception.ProductNotFoundException;
 import com.scaler.productservice.models.Product;
+import com.scaler.productservice.repositories.ProductRepository;
 import com.scaler.productservice.services.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 
@@ -18,16 +18,20 @@ import java.util.List;
 public class ProductController {
 
     private ProductService productService;
+    private ProductRepository productRepository;
+
 
     @Autowired
-    public ProductController(@Qualifier("selfProductService") ProductService productService) {
+    public ProductController(@Qualifier("selfProductService") ProductService productService,
+                             ProductRepository productRepository){
         this.productService = productService;
+        this.productRepository = productRepository;
     }
 
     @GetMapping
-    public ResponseEntity<List<Product>> getAllProducts() {
+    public ResponseEntity<List<Product>> getAllProducts() throws ProductNotFoundException {
         ResponseEntity<List<Product>> response =
-                new ResponseEntity<>(productService.getAllProduct(), HttpStatus.FORBIDDEN);
+                new ResponseEntity<>(productService.getAllProduct(), HttpStatus.OK);
         return response;
     }
 
@@ -54,8 +58,14 @@ public class ProductController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteProduct(@PathVariable("id") Long id) {
+    public void deleteProduct(@PathVariable("id") Long id) throws ProductNotFoundException{
+        if (!productRepository.existsById(id) ){
+            throw new ProductNotFoundException("ider koi product nahi mila " + id + " esi koi id nahi hai");
+        }
         productService.deleteProduct(id);
+        System.out.println("product deleted" + id );
+
+
     }
